remove compute_resources_under_drill to resource-mining
